requirements.txt
Django==5.2.10
djangorestframework==3.16.1
drf-spectacular==0.29.0
